迁移购物车和订单api

+++ Service/CartService.class.php
@@ -15,9 +15,10 @@ class CartService extends BaseService {
      * @param int   $user_id    用户ID
      * @return bool|int|array
      */
-    static function add_cart($goods_id, $goods_num, $goods_spec, $session_id, $user_id = 0) {
-        $goods = M(GoodsService::GOODS_TABLE_NAME)->where("goods_id = $goods_id")->find(); // 找出这个商品
-        $specGoodsPriceList = M('ShopSpecGoodsPrice')->where("goods_id = $goods_id")->getField("key,key_name,price,store_count,sku"); // 获取商品对应的规格价钱 库存 条码
+    static function addCart($goods_id, $goods_num, $goods_spec, $session_id = '', $user_id = 0) {
+        $goods = M(GoodsService::GOODS_TABLE_NAME)->where("goods_id = '%d'", $goods_id)->find(); // 找出这个商品
+        $specGoodsPriceList = M('ShopSpecGoodsPrice')->where("goods_id = '%d'",
+            $goods_id)->getField("key,key_name,price,store_count,sku"); // 获取商品对应的规格价钱 库存 条码
         $user_id = $user_id ? $user_id : 0;
 
         foreach ($goods_spec as $key => $val) {
@@ -29,63 +30,40 @@ class CartService extends BaseService {
             sort($spec_item);
             $spec_key = implode('_', $spec_item);
             if ($specGoodsPriceList[$spec_key]['store_count'] < $goods_num) {
-                return self::createReturn(false,'','商品库存不足');
-//                $this->set_err_msg('商品库存不足');
-//                return false;
+                return self::createReturn(false, '', '商品库存不足');
             }
             $spec_price = $specGoodsPriceList[$spec_key]['price']; // 获取规格指定的价格
         }
 
-        $where = " goods_id = $goods_id and spec_key = '$spec_key' "; // 查询购物车是否已经存在这商品
-        if ($user_id > 0) {
-            $where .= " and (session_id = '$session_id' or user_id = $user_id) ";
-        } else {
-            $where .= " and  session_id = '$session_id' ";
+        // 查询购物车是否已经存在这商品
+        $where = [
+            'goods_id' => $goods_id,
+            'user_id' => $user_id
+        ];
+
+        if ($spec_key) {
+            $where['spec_key'] = $spec_key;
         }
 
         $cart_goods = M(self::TABLE_NAME)->where($where)->find();  // 查找购物车是否已经存在该商品
         $price = $spec_price ? $spec_price : $goods['shop_price']; // 如果商品规格没有指定价格则用商品原始价格
-
-
-        if ($user_id) {
-            $where .= "  or user_id= $user_id ";
-        }
         $catr_count = M(self::TABLE_NAME)->where($where)->count(); // 查找购物车商品总数量
         if ($catr_count >= 20) {
-            return self::createReturn(false,'','购物车最多只能放20种商品');
-
-//            $this->set_err_msg('购物车最多只能放20种商品');
-//
-//            return false;
+            return self::createReturn(false, '', '购物车最多只能放20种商品');
         }
 
         if (!empty($specGoodsPriceList) && empty($goods_spec)) {
             // 有商品规格 但是前台没有传递过来
-            return self::createReturn(false,'','必须传递商品规格');
-//            $this->set_err_msg('必须传递商品规格');
-//
-//            return false;
+            return self::createReturn(false, '', '必须传递商品规格');
         }
         if ($cart_goods['goods_num'] + $goods_num <= 0) {
-            return self::createReturn(false,'','购买商品数量不能为0');
-
-//            $this->set_err_msg('购买商品数量不能为0');
-//
-//            return false;
+            return self::createReturn(false, '', '购买商品数量不能为0');
         }
         if (empty($goods)) {
-            return self::createReturn(false,'','购买商品不存在');
-
-//            $this->set_err_msg('购买商品不存在');
-//
-//            return false;
+            return self::createReturn(false, '', '购买商品不存在');
         }
         if (($goods['store_count'] < ($cart_goods['goods_num'] + $goods_num))) {
-            return self::createReturn(false,'','商品库存不足');
-
-//            $this->set_err_msg('商品库存不足');
-//
-//            return false;
+            return self::createReturn(false, '', '商品库存不足');
         }
 
         $data = array(
@@ -116,29 +94,19 @@ class CartService extends BaseService {
             $update = array();
             $update['goods_num'] = ($cart_goods['goods_num'] + $goods_num);
             $update['add_time'] = time();
-            $res = M(self::TABLE_NAME)->where("id =" . $cart_goods['id'])->save($update); // 数量相加
-            $cart_count = cart_goods_num($user_id, $session_id); // 查找购物车数量
+            $res = M(self::TABLE_NAME)->where("id ='%d'", $cart_goods['id'])->save($update); // 数量相加
+            $cart_count = self::cartGoodsNum($user_id)['data']; // 查找购物车数量
             setcookie('cn', $cart_count, null, '/');
         } else {
-            $res = M(self::TABLE_NAME)->add($data);
-            $cart_count = cart_goods_num($user_id, $session_id); // 查找购物车数量
+            $res = M(self::TABLE_NAME)->add($data)['data'];
+            $cart_count = self::cartGoodsNum($user_id); // 查找购物车数量
             setcookie('cn', $cart_count, null, '/');
         }
         if ($res) {
             //返回购物车id
-            if($cart_goods){
-                return self::createReturn(true,$cart_goods['id'],'添加购物车成功');
-            }else{
-                return self::createReturn(true,$res,'添加购物车成功');
-            }
-//            return $cart_goods ? $cart_goods['id'] : $res;
-
+            return self::createReturn(true, $cart_goods['id'], '添加购物车成功');
         } else {
-            return self::createReturn(false,'','添加购物车失败');
-
-//            $this->set_err_msg('添加购物车失败');
-//
-//            return false;
+            return self::createReturn(false, '', '添加购物车失败');
         }
     }
 
@@ -149,17 +117,17 @@ class CartService extends BaseService {
      * @param int $ids
      * @return array
      */
-    static function get_cart_list($userid,$ids=0){
+    static function getCartList($userid, $ids = 0) {
         //如果用户没有登录，是用session_id加入购物车
         $where = ['user_id' => $userid];
         if ($ids != 0) {
             $where['id'] = ['in', $ids];
         }
         $cart_list = M(CartService::TABLE_NAME)->where($where)->order('id DESC')->select();
-        if($cart_list){
-            return self::createReturn(true,$cart_list,'获取购物车成功!');
-        }else{
-            return self::createReturn(true,[],'购物车为空');
+        if ($cart_list) {
+            return self::createReturn(true, $cart_list, '获取购物车成功!');
+        } else {
+            return self::createReturn(true, [], '购物车为空');
         }
     }
 
@@ -174,17 +142,35 @@ class CartService extends BaseService {
      * 移除购物车
      *
      * @param $cart_id int 购物车id
-     * @param $userid int 用户id
+     * @param $userid  int 用户id
      * @return array
      */
-      static function del_cart($cart_id,$userid){
-          $where['userid'] = $userid;
-          $where['id'] = $cart_id;
-          $res = M(CartService::TABLE_NAME)->where($where)->delete();
-          if ($res) {
-              return self::createReturn(true,$res,'移除购物车成功!');
-          } else {
-              return self::createReturn(false,$res,'无可删除内容!');
-          }
-      }
+    static function delCart($cart_id, $userid) {
+        if ($cart_id == 0) {
+            return self::createReturn(false, '', '参数错误');
+        }
+        $where['userid'] = $userid;
+        $where['id'] = $cart_id;
+        $res = M(CartService::TABLE_NAME)->where($where)->delete();
+        if ($res) {
+            return self::createReturn(true, $res, '移除购物车成功!');
+        } else {
+            return self::createReturn(false, $res, '无可删除内容!');
+        }
+    }
+
+    /**
+     * 查看某个用户购物车中商品的数量
+     *
+     * @param string|int $user_id
+     * @return int 购买数量
+     */
+    static function cartGoodsNum($user_id = 0) {
+        $where['user_id'] = $user_id;
+        // 查找购物车数量
+        $cart_count = M('ShopCart')->where($where)->sum('goods_num');
+        $cart_count = $cart_count ? $cart_count : 0;
+
+        return self::createReturn(true, $cart_count, '');
+    }
 }
